+++ src/util/bit.rs
@@ -14,13 +14,4 @@ pub mod bit {
             panic!("Bit pos out of range (0-7)");
         }
     }
-
-    pub fn get_bit(byte: u8, bit_pos: u8) -> bool {
-        if bit_pos < 8 {
-            byte & (1 << bit_pos) != 0
-        } 
-        else {
-            panic!("Bit pos out of range (0-7)");
-        }
-    }
 }